# tests/unit/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ApplicantsOfJobPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicantOfJobRow"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ApplicantsOfJobPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobCreate": {
        "additionalProperties": false,
        "properties": {
          "ctc_max": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Max"
          },
          "ctc_min": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Min"
          },
          "description": {
            "maxLength": 10000,
            "minLength": 10,
            "title": "Description",
            "type": "string"
          },
          "employer_id": {
            "format": "uuid",
            "title": "Employer Id",
            "type": "string"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Locations",
            "type": "array"
          },
          "max_exp_years": {
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Max Exp Years",
            "type": "integer"
          },
          "min_exp_years": {
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Min Exp Years",
            "type": "integer"
          },
          "status": {
            "default": "open",
            "enum": [
              "open",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "employer_id",
          "title",
          "description",
          "locations",
          "min_exp_years",
          "max_exp_years"
        ],
        "title": "JobCreate",
        "type": "object"
      },
...
      "JobDetailResponse": {
        "properties": {
          "application": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobDetailApplicationRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "employer": {
            "$ref": "#/components/schemas/jobify_api__routes__schemas__EmployerRead"
          },
          "job": {
            "$ref": "#/components/schemas/JobRead"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "saved_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobDetailSavedJobRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "job",
          "employer",
          "match",
          "application",
          "saved_job"
        ],
        "title": "JobDetailResponse",
        "type": "object"
      },
...
      "JobPatch": {
        "additionalProperties": false,
        "properties": {
          "ctc_max": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Max"
          },
          "ctc_min": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Min"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 10000,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "locations": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locations"
          },
          "max_exp_years": {
            "anyOf": [
              {
                "maximum": 50.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Exp Years"
          },
          "min_exp_years": {
            "anyOf": [
              {
                "maximum": 50.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Exp Years"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "open",
                  "closed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "JobPatch",
        "type": "object"
      },
      "JobRead": {
        "properties": {
          "ctc_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Max"
          },
          "ctc_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Min"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "employer_verified": {
            "title": "Employer Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "title": "Locations",
            "type": "array"
          },
          "max_exp_years": {
            "title": "Max Exp Years",
            "type": "integer"
          },
          "min_exp_years": {
            "title": "Min Exp Years",
            "type": "integer"
          },
          "posted_at": {
            "format": "date-time",
            "title": "Posted At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "locations",
          "min_exp_years",
          "max_exp_years",
          "ctc_min",
          "ctc_max",
          "status",
          "posted_at",
          "employer_verified"
        ],
        "title": "JobRead",
        "type": "object"
      },
...
      "RecruiterJobRow": {
        "properties": {
          "applicant_count": {
            "title": "Applicant Count",
            "type": "integer"
          },
          "ctc_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Max"
          },
          "ctc_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Min"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "employer_verified": {
            "title": "Employer Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "title": "Locations",
            "type": "array"
          },
          "max_exp_years": {
            "title": "Max Exp Years",
            "type": "integer"
          },
          "min_exp_years": {
            "title": "Min Exp Years",
            "type": "integer"
          },
          "posted_at": {
            "format": "date-time",
            "title": "Posted At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "surfaced_match_count": {
            "title": "Surfaced Match Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "locations",
          "min_exp_years",
          "max_exp_years",
          "ctc_min",
          "ctc_max",
          "status",
          "posted_at",
          "employer_verified",
          "applicant_count",
          "surfaced_match_count"
        ],
        "title": "RecruiterJobRow",
        "type": "object"
      },
      "RecruiterJobsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecruiterJobRow"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "RecruiterJobsPage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "jobify_api__routes__schemas__EmployerRead": {
        "description": "Wire shape: a verified bool, not the underlying verified_at timestamp.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "verified"
        ],
        "title": "EmployerRead",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Jobify API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs": {
      "post": {
        "operationId": "create_job_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/me": {
      "get": {
        "operationId": "list_my_jobs_v1_jobs_me_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "open",
                    "closed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruiterJobsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_job_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_detail_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Detail",
        "tags": [
          "jobs"
        ]
      },
      "patch": {
        "operationId": "patch_job_v1_jobs__job_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/applicants": {
      "get": {
        "operationId": "list_applicants_for_job_v1_jobs__job_id__applicants_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicantsOfJobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Applicants For Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}